+++ libs/model/src/services/commonProtocol/contractHelpers.ts
@@ -6,7 +6,7 @@ import {
 import Web3 from 'web3';
 import { AbiItem } from 'web3-utils';
 import { DB } from '../../models';
-import { TokenBalanceCache } from '../tokenBalanceCache';
+import { getBalances } from '../tokenBalanceCache';
 
 export const getNamespace = async (
   web3: Web3,
@@ -45,7 +45,6 @@ export const getNamespace = async (
 
 /**
  * gets the balance of an id for an address on a namespace
- * @param tbc TokenBalanceCache instance
  * @param namespace namespace name
  * @param tokenId ERC1155 id(ie 0 for admin token, default 2 for CommunityStake)
  * @param chain chainNode to use(must be chain with deployed protocol)
@@ -54,7 +53,6 @@ export const getNamespace = async (
  * @returns balance in wei
  */
 export const getNamespaceBalance = async (
-  tbc: TokenBalanceCache,
   namespace: string,
   tokenId: number,
   chain: commonProtocol.ValidChains,
@@ -78,7 +76,7 @@ export const getNamespaceBalance = async (
     if (activeNamespace === '0x0000000000000000000000000000000000000000') {
       throw new AppError('Namespace not found for this name');
     }
-    const balance = await tbc.getBalances({
+    const balance = await getBalances({
       balanceSourceType: BalanceSourceType.ERC1155,
       addresses: [address],
       sourceOptions: {
