+++ libs/model/src/services/index.ts
@@ -1,2 +1,3 @@
 export * as commonProtocol from './commonProtocol';
 export * as tokenBalanceCache from './tokenBalanceCache';
+export * from './tokenBalanceCache/types';
