+++ packages/commonwealth/server/controllers/server_comments_controller.ts
@@ -1,7 +1,6 @@
 import { DB } from '@hicommonwealth/model';
 import BanCache from '../util/banCheckCache';
 
-import { TokenBalanceCache } from '@hicommonwealth/model';
 import GlobalActivityCache from '../util/globalActivityCache';
 
 import {
@@ -32,7 +31,6 @@ import {
 export class ServerCommentsController {
   constructor(
     public models: DB,
-    public tokenBalanceCache: TokenBalanceCache,
     public banCache: BanCache,
     public globalActivityCache?: GlobalActivityCache,
   ) {}
