+++ packages/commonwealth/server/controllers/server_comments_methods/create_comment_reaction.ts
@@ -8,7 +8,7 @@ import {
   AddressInstance,
   ReactionAttributes,
   UserInstance,
-  contractHelpers,
+  commonProtocol as commonProtocolService,
 } from '@hicommonwealth/model';
 import { REACTION_WEIGHT_OVERRIDE } from 'server/config';
 import { MixpanelCommunityInteractionEvent } from '../../../shared/analytics/types';
@@ -96,7 +96,6 @@ export async function __createCommentReaction(
     try {
       const { isValid } = await validateTopicGroupsMembership(
         this.models,
-        this.tokenBalanceCache,
         thread.topic_id,
         thread.community_id,
         address,
@@ -126,14 +125,14 @@ export async function __createCommentReaction(
       if (!community) {
         throw new AppError(Errors.CommunityNotFound);
       }
-      const stakeBalance = await contractHelpers.getNamespaceBalance(
-        this.tokenBalanceCache,
-        community.namespace,
-        stake.stake_id,
-        commonProtocol.ValidChains.Sepolia,
-        address.address,
-        this.models,
-      );
+      const stakeBalance =
+        await commonProtocolService.contractHelpers.getNamespaceBalance(
+          community.namespace,
+          stake.stake_id,
+          commonProtocol.ValidChains.Sepolia,
+          address.address,
+          this.models,
+        );
       calculatedVotingWeight = commonProtocol.calculateVoteWeight(
         stakeBalance,
         voteWeight,
