+++ packages/commonwealth/server/controllers/server_groups_methods/refresh_membership.ts
@@ -45,7 +45,6 @@ export async function __refreshMembership(
 
   const memberships = await refreshMembershipsForAddress(
     this.models,
-    this.tokenBalanceCache,
     address,
     groups,
     true, // use fresh balances
