+++ packages/commonwealth/server/controllers/server_threads_methods/create_thread_comment.ts
@@ -136,7 +136,6 @@ export async function __createThreadComment(
   if (!isAdmin) {
     const { isValid, message } = await validateTopicGroupsMembership(
       this.models,
-      this.tokenBalanceCache,
       thread.topic_id,
       thread.community_id,
       address,
