+++ packages/commonwealth/test/integration/api/getRelatedCommunities.spec.ts
@@ -8,7 +8,7 @@ describe('GetRelatedCommunities Tests', () => {
   });
 
   it('Correctly returns nothing if base does not match chainNode', async () => {
-    const controller = new ServerCommunitiesController(models, null, null);
+    const controller = new ServerCommunitiesController(models, null);
     const response = await controller.getRelatedCommunities({
       chainNodeId: -100,
     });
@@ -17,7 +17,7 @@ describe('GetRelatedCommunities Tests', () => {
   });
 
   it('Correctly returns results if base matches some chainNode.name', async () => {
-    const controller = new ServerCommunitiesController(models, null, null);
+    const controller = new ServerCommunitiesController(models, null);
     const response = await controller.getRelatedCommunities({ chainNodeId: 2 });
 
     assert.equal(response.length, 3);
