+++ packages/commonwealth/test/unit/server_controllers/server_comments_controller.spec.ts
@@ -1,5 +1,5 @@
 import { NotificationCategories } from '@hicommonwealth/core';
-import { CommunityInstance, contractHelpers } from '@hicommonwealth/model';
+import { CommunityInstance, commonProtocol } from '@hicommonwealth/model';
 import { expect } from 'chai';
 import { ServerCommentsController } from 'server/controllers/server_comments_controller';
 import { SearchCommentsOptions } from 'server/controllers/server_comments_methods/search_comments';
@@ -8,7 +8,9 @@ import { BAN_CACHE_MOCK_FN } from 'test/util/banCacheMock';
 
 describe('ServerCommentsController', () => {
   beforeEach(() => {
-    Sinon.stub(contractHelpers, 'getNamespaceBalance').resolves('0');
+    Sinon.stub(commonProtocol.contractHelpers, 'getNamespaceBalance').resolves(
+      '0',
+    );
   });
   afterEach(() => {
     Sinon.restore();
@@ -75,7 +77,6 @@ describe('ServerCommentsController', () => {
           }),
         },
       };
-      const tokenBalanceCache = {};
       const banCache = BAN_CACHE_MOCK_FN('ethereum');
 
       const user = {
@@ -91,7 +92,6 @@ describe('ServerCommentsController', () => {
 
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
 
@@ -185,7 +185,6 @@ describe('ServerCommentsController', () => {
           }),
         },
       };
-      const tokenBalanceCache = {};
       const banCache = {
         checkBan: sandbox.stub().resolves([true, null]),
       };
@@ -198,7 +197,6 @@ describe('ServerCommentsController', () => {
 
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
 
@@ -250,7 +248,6 @@ describe('ServerCommentsController', () => {
           findOne: sandbox.stub().resolves(null),
         },
       };
-      const tokenBalanceCache = {};
       const banCache = {
         checkBan: sandbox.stub().resolves([true, null]),
       };
@@ -263,7 +260,6 @@ describe('ServerCommentsController', () => {
 
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
 
@@ -317,7 +313,6 @@ describe('ServerCommentsController', () => {
           }),
         },
       };
-      const tokenBalanceCache = {};
       const banCache = {
         checkBan: sandbox.stub().resolves([false, 'big ban err']),
       };
@@ -331,7 +326,6 @@ describe('ServerCommentsController', () => {
 
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
 
@@ -442,9 +436,6 @@ describe('ServerCommentsController', () => {
           bulkCreate: sandbox.stub().resolves([]),
         },
       };
-      const tokenBalanceCache = {
-        getBalances: sandbox.stub().resolves({}),
-      };
       const banCache = {
         checkBan: sandbox.stub().resolves([true, null]),
       };
@@ -460,7 +451,6 @@ describe('ServerCommentsController', () => {
 
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
 
@@ -489,12 +479,10 @@ describe('ServerCommentsController', () => {
           },
         },
       };
-      const tokenBalanceCache = {};
       const banCache = {};
 
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
 
@@ -553,13 +541,11 @@ describe('ServerCommentsController', () => {
           findOne: async () => data,
         },
       };
-      const tokenBalanceCache = {};
       const banCache = {
         checkBan: async () => [true, null],
       };
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
       const user = {
@@ -631,11 +617,9 @@ describe('ServerCommentsController', () => {
           findOne: async () => data,
         },
       };
-      const tokenBalanceCache = {};
       const banCache = BAN_CACHE_MOCK_FN('ethereum');
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
       const user = {
@@ -682,13 +666,11 @@ describe('ServerCommentsController', () => {
           findOne: async () => null,
         },
       };
-      const tokenBalanceCache = {};
       const banCache = {
         checkBan: async () => [true, null],
       };
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
       const user = {
@@ -732,14 +714,12 @@ describe('ServerCommentsController', () => {
           destroy: async () => ({}),
         },
       };
-      const tokenBalanceCache = {};
       const banCache = {
         checkBan: () => [true, null],
       };
 
       const serverCommentsController = new ServerCommentsController(
         db as any,
-        tokenBalanceCache as any,
         banCache as any,
       );
