+++ packages/commonwealth/test/unit/server_controllers/server_reactions_controller.spec.ts
@@ -1,4 +1,4 @@
-import { contractHelpers } from '@hicommonwealth/model';
+import { commonProtocol } from '@hicommonwealth/model';
 import { expect } from 'chai';
 import { ServerReactionsController } from 'server/controllers/server_reactions_controller';
 import Sinon from 'sinon';
@@ -6,7 +6,9 @@ import { BAN_CACHE_MOCK_FN } from 'test/util/banCacheMock';
 
 describe('ServerReactionsController', () => {
   beforeEach(() => {
-    Sinon.stub(contractHelpers, 'getNamespaceBalance').resolves('0');
+    Sinon.stub(commonProtocol.contractHelpers, 'getNamespaceBalance').resolves(
+      '0',
+    );
   });
   afterEach(() => {
     Sinon.restore();
